+++ internal/applications/filesharer/internal/handler/getters.go
@@ -2,15 +2,18 @@ package handler
 
 import (
 	"github.com/number571/hidden-lake/internal/applications/filesharer/pkg/app/config"
+	"github.com/number571/hidden-lake/internal/applications/filesharer/pkg/settings"
 	"github.com/number571/hidden-lake/internal/utils/language"
 )
 
 type sTemplate struct {
+	FAppName  string
 	FLanguage language.ILanguage
 }
 
 func getTemplate(pCfg config.IConfig) *sTemplate {
 	return &sTemplate{
+		FAppName:  settings.CServiceName,
 		FLanguage: pCfg.GetSettings().GetLanguage(),
 	}
 }