+++ internal/applications/messenger/pkg/app/init_service.go
@@ -11,8 +11,8 @@ import (
 	"github.com/number571/hidden-lake/internal/applications/messenger/internal/msgbroker"
 	"github.com/number571/hidden-lake/internal/applications/messenger/pkg/app/config"
 	hlm_settings "github.com/number571/hidden-lake/internal/applications/messenger/pkg/settings"
-	"github.com/number571/hidden-lake/internal/applications/messenger/web"
 	hls_client "github.com/number571/hidden-lake/internal/service/pkg/client"
+	"github.com/number571/hidden-lake/internal/webui"
 	"golang.org/x/net/websocket"
 )
 
@@ -46,7 +46,7 @@ func (p *sApp) initInterfaceServiceHTTP(
 	mux := http.NewServeMux()
 	mux.Handle(hlm_settings.CStaticPath, http.StripPrefix(
 		hlm_settings.CStaticPath,
-		handleFileServer(p.fHTTPLogger, p.fConfig, http.FS(web.GetStaticPath()))),
+		handleFileServer(p.fHTTPLogger, p.fConfig, http.FS(webui.GetStaticPath()))),
 	)
 
 	cfgWrapper := config.NewWrapper(p.fConfig)