+++ internal/webui/web_test.go
@@ -1,0 +1,10 @@
+package webui
+
+import (
+	"testing"
+)
+
+func TestPath(_ *testing.T) {
+	_ = GetTemplatePath()
+	_ = GetStaticPath()
+}